+++ tools/str_gen/xml_str_gen.cc
@@ -99,6 +99,8 @@ static void xml_gen_on_comment(XmlBuilder* thiz, const char* text, size_t length
 }
 
 static XmlBuilder* builder_init(xml_str_builder_t& b, StrGen* sg) {
+  memset(&b, 0x00, sizeof(xml_str_builder_t));
+
   b.sg = sg;
   b.level = 0;
   b.builder.on_start = xml_str_gen_on_start;